Fixes dependencies and enables debugging
It turns out that with gradle one needs Thymeleaf as explicit dependency. So this
is available now, including the Layout dialect. Additionally, `home.html` is now
having a Thymeleaf compatible format (it was compaining about a non-closed
`<meta>` tag...).
Then, log level `DEBUG` is set for everything under the package `mk2`.