LDAP configuration works
With this commit, the authentication is using properties prefixed with
`mk2.ldap.contextSource` - check `application.properties`.

The default config points to a local LDAP listening on port 389 and
using the same structure as the mehlbox is doing at the moment. In
addition to roles derived by group memberships, the role "ROLE_USER" is
granted to all users who authenticate successful. This is useful to give
basic permissions without the need of adding all users to a dedicated
LDAP group.

Last but not least: `home` shows a comma-separated list of roles the
currently authenticated user has.