Delete mail addresses
This commit adds delete functionality for email addresses assigned to
the currently authenticated user.
It brings in a really ugly hack, as we can not use DELETE for the end-
point (not supported by HTML). Also `EmailAddressLogic` isn't
necessarily my favourite so far. But it does what it is supposed to do.