Merge new version 0.3.1 (#13)
* IDEA changes.

* Rebase merge

* Fixing Gradle build issues.

* Update to Spring Boot 2 (#6)

* Updating License Block and Spring Boot & related changes

* Incrementing version

* IntelliJ foo

* Fixed Change Password Page & Rendering Issues

* Updating My Domains View

* Aligns and styles "Change" button

* Get's rid of redundant copyright notes

* Again Removes Duplicate Copyright

* Adds release make target

* Some changes from the import

* Update dependencies and clean up (#12)

* WIP

* Updating dependencies and adding meta information

This will bring all the dependencies to the latest version.
Also, the log in screen and the change password dialog are using
the `autocomplete` attribute for better password manager support now.

* Removes webjars

Webjars weren't used anyways. This commit will remove the dependencies.

* Bumps version to 0.3.0

* Bumps version to 0.3.1

Essentially a clean up and merge of the latest dependency updates and develop.