Chore: Upgrade versions
There were changes in Spring Security that deprecated Web Security related things. This also led to a circular dependency issues during Spring's startup time.

This fixes it at least to a point where the application appears to be starting.

Please check the FIXME comment and this documentation: https://spring.io/blog/2022/02/21/spring-security-without-the-websecurityconfigureradapter